Move ReflectanceCorrector to satpy.modifiers
