+++ satpy/etc/composites/abi.yaml
@@ -2,7 +2,7 @@ sensor_name: visir/abi
 
 modifiers:
   rayleigh_corrected_crefl:
-    modifier: !!python/name:satpy.composites.viirs.ReflectanceCorrector
+    modifier: !!python/name:satpy.modifiers.atmosphere.ReflectanceCorrector
     dem_filename: CMGDEM.hdf
     optional_prerequisites:
       - name: satellite_azimuth_angle
